refactor: remove deprecated components
